fbt: moved main code to scripts folder
