+++ scripts/fbt/appmanifest.py
