+++ scripts/fbt/elfmanifest.py
