+++ scripts/fbt/util.py
@@ -41,25 +41,3 @@ def link_dir(target_path, source_path, is_windows):
 
 def single_quote(arg_list):
     return " ".join(f"'{arg}'" if " " in arg else str(arg) for arg in arg_list)
-
-
-def link_elf_dir_as_latest(env, elf_node):
-    elf_dir = elf_node.Dir(".")
-    latest_dir = env.Dir("#build/latest")
-    print(f"Setting {elf_dir} as latest built dir (./build/latest/)")
-    return link_dir(latest_dir.abspath, elf_dir.abspath, env["PLATFORM"] == "win32")
-
-
-def should_gen_cdb_and_link_dir(env, requested_targets):
-    explicitly_building_updater = False
-    # Hacky way to check if updater-related targets were requested
-    for build_target in requested_targets:
-        if "updater" in str(build_target):
-            explicitly_building_updater = True
-
-    is_updater = not env["IS_BASE_FIRMWARE"]
-    # If updater is explicitly requested, link to the latest updater
-    # Otherwise, link to firmware
-    return (is_updater and explicitly_building_updater) or (
-        not is_updater and not explicitly_building_updater
-    )
