+++ scripts/fbt/version.py
