+++ scripts/fbt_tools/blackmagic.py
