+++ scripts/fbt_tools/fbt_apps.py
