+++ scripts/fbt_tools/fbt_assets.py
