+++ scripts/fbt_tools/fbt_help.py
