+++ scripts/fbt_tools/fbt_version.py
