+++ scripts/fbt_tools/objdump.py
