+++ scripts/fbt_tools/openocd.py
