+++ scripts/fbt_tools/strip.py
